+++ redun/executors/aws_batch.py
@@ -3,6 +3,7 @@ import datetime
 import json
 import logging
 import re
+import subprocess
 import threading
 import time
 import uuid
@@ -65,6 +66,47 @@ ARRAY_JOB_SUFFIX = "array"
 DOCKER_INSPECT_ERROR = "CannotInspectContainerError: Could not transition to inspecting"
 BATCH_JOB_TIMEOUT_ERROR = "Job attempt duration exceeded timeout"
 DEBUG_SCRATCH = "scratch"
+
+
+def get_default_registry() -> str:
+    """
+    Returns the default ECR registry.
+    """
+    client = boto3.client("ecr")
+    resp = client.get_authorization_token()
+    registry = resp["authorizationData"][0]["proxyEndpoint"].strip("https://")
+    return registry
+
+
+def get_or_create_repo(repo_name: str) -> dict:
+    """
+    Get or create an ECR repository.
+    """
+    client = boto3.client("ecr")
+
+    resp = client.describe_repositories(repositoryNames=[repo_name])
+    if resp["repositories"]:
+        repo = resp["repositories"][0]
+    else:
+        repo = client.create_repository(repositoryName=repo_name)["repository"]
+    return repo
+
+
+def ecr_push(repo_name: str, registry: Optional[str] = None) -> None:
+
+    # Get default registry based on login.
+    if registry is None:
+        registry = get_default_registry()
+
+    get_or_create_repo(repo_name)
+
+    # docker uri
+    ecr_uri = "/".join([registry, repo_name])
+
+    subprocess.check_call(["docker", "build", "-t", ecr_uri, "."])
+
+    # push
+    subprocess.check_call(["docker", "push", ecr_uri])
 
 
 def is_array_job_name(job_name: str) -> bool:
@@ -749,6 +791,7 @@ class AWSBatchExecutor(Executor):
         self._docker_executor.set_scheduler(scheduler)
 
     def gather_inflight_jobs(self) -> None:
+
         running_arrays: Dict[str, List[Tuple[str, int]]] = defaultdict(list)
 
         # Get all running jobs by name
