--- redun/executors/s3_utils.py
@@ -1,7 +1,0 @@
-# S3 scratch filenames.
-S3_SCRATCH_INPUT = "input"
-S3_SCRATCH_OUTPUT = "output"
-S3_SCRATCH_CODE = "code.tar.gz"
-S3_SCRATCH_ERROR = "error"
-S3_SCRATCH_HASHES = "eval_hashes"
-S3_SCRATCH_STATUS = "status"
